Extract the palette driver the /sleep test needs
Moves the overlay-page lookup and command typing into a shared support
helper so more scenarios can drive commands the way a person does. The
helper waits on the RENDERER's overlay mode, not main's — main flips
first, and filling the input on that signal races it into existence.

Found while auditing for the same shape as the /sleep bug: the generic
"I run the slash command" step in runnable.steps.js switches on the
command name and calls the function behind it, so no scenario has ever
typed a slash command. Converting it surfaced two further problems — a
find-bar assertion that does not wait for the IPC hop it now depends on,
and an intermittent overlay-visibility race in the ad-blocking scenarios.
Both want their own investigation, so that conversion is deliberately not
in this commit rather than landing an intermittently-red suite.
