+++ test/desktop/steps/quiet-tabs.steps.js
@@ -1,24 +1,8 @@
 const assert = require('node:assert/strict');
 const { Given, When, Then } = require('@cucumber/cucumber');
 const { waitForValue } = require('../support/poll');
-const ctx = require('../support/context');
+const { runSlashCommand } = require('../support/overlay');
 
-// The overlay is its own page, so the panel can be driven the way a person
-// drives it. Calling the main-process function instead is what let the
-// /sleep receipt regress unnoticed: the command quieted tabs and left the
-// list showing slash commands, so nothing on screen changed.
-async function overlayPage() {
-  for (let attempt = 0; attempt < 40; attempt += 1) {
-    const page = ctx.app.windows().find((candidate) =>
-      !candidate.isClosed() && candidate.url().endsWith('/src/renderer/overlay.html'));
-    if (page) {
-      await page.waitForLoadState('domcontentloaded');
-      return page;
-    }
-    await new Promise((resolve) => setTimeout(resolve, 100));
-  }
-  throw new Error('overlay window never appeared');
-}
 
 async function openQuietable(world, name, opts = {}) {
   const suffix = opts.extraQuery ? `&${opts.extraQuery}` : '';
@@ -191,9 +175,7 @@ When('I run the manual sleep command', async function () {
   // Type it and press Enter, exactly as a person would. Driving
   // sleepBackgroundTabsNow directly leaves the input empty, which is the one
   // state in which the panel happens to be showing the tab rows.
-  const page = await overlayPage();
-  await page.fill('#addressInput', '/sleep');
-  await page.press('#addressInput', 'Enter');
+  await runSlashCommand(this, '/sleep');
   await this.waitForState((state) =>
     state.tabs.find((candidate) => candidate.id === this.quietCandidateId)?.asleep === true);
 });
